+++ src/utils/mod.rs
@@ -1,8 +1,7 @@
 pub mod file;
-mod merge;
+pub mod or_else_try;
 pub mod string;
 mod try_convert;
 
 pub use file::get_file_contents;
-pub use merge::merge;
 pub use try_convert::TryConvert;
