+++ src/utils/string.rs
@@ -21,6 +21,16 @@ impl OptionStr<String> for Option<String> {
     }
 }
 
+impl OptionStr<String> for String {
+    fn none_if_empty(self) -> Option<String> {
+        if self.is_empty() {
+            None
+        } else {
+            Some(self)
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
